+++ src/ContactInformation/ContactInfoFormGroup/EmailAddresses.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Field } from 'redux-form';
-import { Row, Col, Button, TextField, Select } from '@folio/stripes-components';
+import { Field, getFormValues } from 'redux-form';
+import { MultiSelection, Row, Col, Button, TextField, Select } from '@folio/stripes-components';
 import css from '../ContactInfoFormGroup.css';
 import { Required } from '../../Utils/Validate';
 
@@ -9,12 +9,31 @@ class EmailAddresses extends Component {
   static propTypes = {
     dropdownCategories: PropTypes.arrayOf(PropTypes.object),
     dropdownLanguages: PropTypes.arrayOf(PropTypes.object),
-    fields: PropTypes.object
+    fields: PropTypes.object,
+    stripes: PropTypes.shape({
+      store: PropTypes.func
+    }),
+    dispatch: PropTypes.func,
+    change: PropTypes.func,
   };
 
   constructor(props) {
     super(props);
     this.renderSubEmailAddresses = this.renderSubEmailAddresses.bind(this);
+    this.onChangeSelect = this.onChangeSelect.bind(this);
+    this.selectedValues = this.selectedValues.bind(this);
+  }
+
+  onChangeSelect = (e, elem, propertyName) => {
+    const { dispatch, change } = this.props;
+    dispatch(change(`${elem}.${propertyName}`, e));
+  }
+
+  selectedValues = (index, fields, propertyName) => {
+    const { stripes: { store } } = this.props;
+    const formValues = getFormValues('FormVendor')(store.getState());
+    const currValues = formValues[fields.name][index][propertyName];
+    return currValues;
   }
 
   renderSubEmailAddresses = (elem, index, fields) => {
@@ -31,7 +50,7 @@ class EmailAddresses extends Component {
           <Field label="Default Language" name={`${elem}.language`} id={`${elem}.language`} component={Select} fullWidth dataOptions={dropdownLanguages} />
         </Col>
         <Col xs={12} md={3}>
-          <Field label="Category" name={`${elem}.categories`} id={`${elem}.categories`} component={Select} fullWidth dataOptions={dropdownCategories} style={{ height: '80px' }} multiple />
+          <MultiSelection label="Categories" name={`${elem}.categories`} dataOptions={dropdownCategories} onChange={(e) => this.onChangeSelect(e, elem, 'categories')} style={{ height: '80px' }} value={this.selectedValues(index, fields, 'categories')} />
         </Col>
         <Col xs={12} md={3} mdOffset={9} style={{ textAlign: 'right' }}>
           <Button onClick={() => fields.remove(index)} buttonStyle="danger">