+++ src/ContactInformation/ContactInfoFormGroup/Urls.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { Field } from 'redux-form';
-import { Row, Col, Button, TextField, Select } from '@folio/stripes-components';
+import { Field, getFormValues } from 'redux-form';
+import { MultiSelection, Row, Col, Button, TextField, Select } from '@folio/stripes-components';
 import css from '../ContactInfoFormGroup.css';
 import { Required } from '../../Utils/Validate';
 
@@ -9,12 +9,31 @@ class Url extends Component {
   static propTypes = {
     dropdownCategories: PropTypes.arrayOf(PropTypes.object),
     dropdownLanguages: PropTypes.arrayOf(PropTypes.object),
-    fields: PropTypes.object
+    fields: PropTypes.object,
+    stripes: PropTypes.shape({
+      store: PropTypes.func
+    }),
+    dispatch: PropTypes.func,
+    change: PropTypes.func,
   };
 
   constructor(props) {
     super(props);
     this.renderSubUrl = this.renderSubUrl.bind(this);
+    this.onChangeSelect = this.onChangeSelect.bind(this);
+    this.selectedValues = this.selectedValues.bind(this);
+  }
+
+  onChangeSelect = (e, elem, propertyName) => {
+    const { dispatch, change } = this.props;
+    dispatch(change(`${elem}.${propertyName}`, e));
+  }
+
+  selectedValues = (index, fields, propertyName) => {
+    const { stripes: { store } } = this.props;
+    const formValues = getFormValues('FormVendor')(store.getState());
+    const currValues = formValues[fields.name][index][propertyName];
+    return currValues;
   }
 
   renderSubUrl = (elem, index, fields) => {
@@ -28,7 +47,7 @@ class Url extends Component {
           <Field label="Description" name={`${elem}.url.description`} id={`${elem}.url.description`} component={TextField} fullWidth />
         </Col>
         <Col xs={12} md={3}>
-          <Field label="Categories" name={`${elem}.categories`} id={`${elem}.categories`} component={Select} dataOptions={dropdownCategories} style={{ height: '80px' }} fullWidth multiple />
+          <MultiSelection label="Categories" name={`${elem}.categories`} dataOptions={dropdownCategories} onChange={(e) => this.onChangeSelect(e, elem, 'categories')} style={{ height: '80px' }} value={this.selectedValues(index, fields, 'categories')} />
         </Col>
         <Col xs={12} md={3}>
           <Field label="Default Language" name={`${elem}.language`} id={`${elem}.language`} component={Select} fullWidth dataOptions={dropdownLanguages} />